refactor: extract all inline test modules to companion *_tests.rs files
Moves every #[cfg(test)] mod tests { ... } block out of source files
into a dedicated companion file, following the pattern already used by
indexer.rs, parser.rs, rg.rs, and the indexer/ submodules.

Each source file now has only:
  #[cfg(test)]
  #[path = "foo_tests.rs"]
  mod tests;

Files migrated:
  src/backend/{handlers,helpers,rename}_tests.rs
  src/indexer/{node_ext,resolution}_tests.rs
  src/{inlay_hints,lines_ext,str_ext}_tests.rs
  src/resolver/infer_tests.rs

675 tests pass. No production code changed.
